made itemsForGoal an int at the source

+++ worlds/pd2_crimdawn/world.py
@@ -42,7 +42,7 @@ class CrimDawnWorld(World):
         self.maxTimeBonuses = round((self.options.run_length.value * 15) / self.options.progression_pacing.value) - 1
         # run_length * 15 = 60 (4 heists) or 90 (6 heists)
         # (minutes / pacing) - 1 = items needed to hit run_length
-        self.itemsForGoal = (self.options.run_length.value * 10) / self.options.progression_pacing.value - 1
+        self.itemsForGoal = int((self.options.run_length.value * 10) / self.options.progression_pacing.value - 1)
 
         if self.options.biglobby == 0: self.botCount = 3
         else: self.botCount = self.random.randint(4,21)
